+++ include/FREDServer/Parser/utility.h
@@ -10,15 +10,15 @@ using namespace std;
 class Utility
 {
 public:
-    static vector<uint32_t> splitString2Num(const string& text, string by);
+    static vector<double> splitString2Num(const string& text, string by);
     static vector<string> splitString(const string& text, string by);
     static vector<vector<string> > splitMessage(const string& text);
-    static vector<vector<uint32_t> > splitMessage2Num(const string& text);
+    static vector<vector<double> > splitMessage2Num(const string& text);
     static void removeWhiteSpaces(string& text);
     static void removeComment(string& text);
     static void printVector(const vector<string>& data);
     static void printVector(const vector<int32_t>& data);
-    static double calculateEquation(string& equation, vector<string>& variables, vector<uint32_t>& values);
+    static double calculateEquation(string& equation, vector<string>& variables, vector<double>& values);
     static string readbackToString(vector<double> data);
     static vector<unsigned long> splitAlfResponse(const string &message, Instructions::Type type);
     static void checkMessageIntegrity(const string& request, const string& response, Instructions::Type type);
